+++ src/main/java/dandi/dandi/post/web/PostControllerDocs.java
@@ -24,6 +24,7 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -98,4 +99,12 @@ public interface PostControllerDocs {
                                                                           Double minTemperature,
                                                                           @RequestParam(value = "max")
                                                                           Double maxTemperature);
+
+    @Operation(summary = "게시글 신고")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "201", description = "게시글 신고 성공"),
+            @ApiResponse(responseCode = "404", description = "존재하지 않는 게시글"),
+            @ApiResponse(responseCode = "400", description = "이미 신고한 게시글"),
+    })
+    ResponseEntity<Void> reportPost(@Parameter(hidden = true) Long memberId, @PathVariable Long postId);
 }
