--- src/main/java/dandi/dandi/postreport/application/port/in/PostReportUseCase.java
@@ -1,6 +1,0 @@
-package dandi.dandi.postreport.application.port.in;
-
-public interface PostReportUseCase {
-
-    void reportPost(Long memberId, Long postId);
-}
